Prefer httpx.Auth instead of api_token
It is still possible to use the api_token parameter.
In this case we assume an API token was grabbed from the
user profile of a Dataverse instance.

If both are specified, an api_token and an explicit auth method,
we will warn the user and use the auth method.

Closes #192.